Update girls.txt
What about Florence (great-great grandmother's name on maternal side)? Or Anina or Maria (but pronounced Mar eye a, as mine should have been) !!!!
Well, it was fun looking at them all....... as you know, I like names that stand the test of time - Nathan has 3 really good names. Well, I think so.... but it hasn't got anything to do with me! Meena, Rowena (now the latter one I really like!!), Seren (Welsh for 'star' but is also a girl's name)............... I could suggest names till the cows come home...... but really it's up to you guys, isn't it, but I beg you by all that is holy, not Ada!!!!! The poor kid to be landed with THAT!!!!!!